rolls back to preshell check. adds solarize check

+++ scripts/apps.sh
@@ -41,12 +41,16 @@ lend_me_your "Sublime Text 2" "http://c758482.r82.cf2.rackcdn.com/Sublime%20Text
 # lend_me_your "Sublime Text"   "http://c758482.r82.cf2.rackcdn.com/Sublime%20Text%20Build%203059.dmg"
 
 # Solarized Terminal colors
-echo "Installing 'Solarized Colors' for Terminal..."
-curl -L -o "solarized.zip" https://github.com/tomislav/osx-terminal.app-colors-solarized/archive/master.zip
-unzip "solarized.zip"
-open "osx-terminal.app-colors-solarized-master/Solarized Light.terminal"
-open "osx-terminal.app-colors-solarized-master/Solarized Dark.terminal"
-rm solarized.zip
-defaults write com.apple.terminal "Default Window Settings" -string "Solarized Dark"
-defaults write com.apple.terminal "Startup Window Settings" -string "Solarized Dark"
-rm -rf "osx-terminal.app-colors-solarized-master"
+if [[ $term_default != *Solarized* ]]; then
+  echo "Installing 'Solarized Colors' for Terminal..."
+  curl -L -o "solarized.zip" https://github.com/tomislav/osx-terminal.app-colors-solarized/archive/master.zip;
+  unzip "solarized.zip"
+  open "osx-terminal.app-colors-solarized-master/Solarized Light.terminal"
+  open "osx-terminal.app-colors-solarized-master/Solarized Dark.terminal"
+  rm solarized.zip
+  defaults write com.apple.terminal "Default Window Settings" -string "Solarized Dark"
+  defaults write com.apple.terminal "Startup Window Settings" -string "Solarized Dark"
+  rm -rf "osx-terminal.app-colors-solarized-master"
+else
+  echo "already installed";
+fi
